Make scrums identifiable
ForEach structures generate dynamic views by iterating over identifiable data. In the previous commit, a key path was used to identify scrums by their titles. This approach worked for prototyping because each scrum in the sample data has a unique title. But users would experience issues if they tried to create a new scrum with the same title as an existing one.
To work with user-generated content, changes in this commit make DailyScrum conform to the Identifiable protocol for a more robust way to communicate identity. The protocol has one requirement: an id property that provides a stable identifier for the entity.