Add a new data structure, Data, to the Data Model
Create a new Data type inside DailyScrum. Data will contain all the editable properties of DailyScrum. The types of the properties match the controls that will be created in the edit view.

As the user edits, the view stores the scrum values in a Data property. If the user confirms the change, the view updates DailyScrum using the contents of Data.